+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 1.0.0
 
 - Create template project and basic models [ISSUE-1](https://github.com/kiranabazaar7/kirana-bazaar-js/issues/1)
+- Create landing UI for login [ISSUE-3](https://github.com/kiranabazaar7/kirana-bazaar-js/issues/3)
+- Create UI from Admin login screen [ISSUE-4](https://github.com/kiranabazaar7/kirana-bazaar-js/issues/4)
+- Create UI for Store header [ISSUE-5](https://github.com/kiranabazaar7/kirana-bazaar-js/issues/5)